requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-aiplatform
pandas
db-dtypes
flask
flask-cors
